requirements.txt
absl-py==1.4.0
addict==2.4.0
adjustText==1.2.0
aiohttp==3.8.4
aiosignal==1.3.1
ansi2html==1.8.0
antlr4-python3-runtime==4.9.3
anyio==3.7.0
apptools==5.2.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.3.0
bleach==6.0.0
cachetools==5.3.1
ccimport==0.4.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
colorama==0.4.6
conda-pack==0.8.0
ConfigArgParse==1.5.5
configobj==5.0.8
correlation-cuda==0.0.0
cumm==0.4.11
cumm-cu113==0.4.11
cycler==0.11.0
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
descartes==1.1.0
easydict==1.10
einops==0.6.1
entrypoints==0.4
envisage==7.0.3
exceptiongroup==1.1.1
fastjsonschema==2.17.1
filelock==3.12.0
fire==0.5.0
flake8==3.9.2
Flask==2.2.5
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
future==0.18.3
fvcore==0.1.5.post20221221
google-auth==2.19.1
google-auth-oauthlib==0.4.6
grpcio==1.54.2
huggingface-hub==0.14.1
hydra-core==1.3.2
idna==3.4
imageio==2.31.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==2.0.0
iopath==0.1.10
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-server==1.24.0
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
kornia==0.6.12
lark==1.1.8
llvmlite==0.31.0
lpips==0.1.4
lyft-dataset-sdk==0.0.8
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.5.2
matplotlib-inline==0.1.6
mayavi==4.8.1
mccabe==0.6.1
mdurl==0.1.2
mistune==2.0.5
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mmcv==1.3.8
mmcv-full==1.4.0
mmdet==2.14.0
mmengine==0.7.4
mmsegmentation==0.14.1
model-index==0.1.11
monai==1.1.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.5.0
nest-asyncio==1.5.6
networkx==2.2
ninja==1.11.1.1
notebook==6.5.4
notebook_shim==0.2.3
numba==0.48.0
numpy==1.19.5
nuscenes-devkit==1.1.10
oauthlib==3.2.2
omegaconf==2.3.0
open3d==0.16.0
open3d-python==0.7.0.0
opencv-python==4.7.0.72
openmim==0.3.7
ordered-set==4.1.0
packaging==23.1
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pccm==0.4.11
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1660385854171/work
pkgutil_resolve_name==1.3.10
platformdirs==3.5.1
plotly==5.14.1
pluggy==1.0.0
plyfile==0.9
portalocker==2.7.0
prettytable==3.7.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pycocotools==2.0.6
pycodestyle==2.7.0
pycparser==2.21
pyDeprecate==0.3.0
pyface==8.0.0
pyflakes==2.3.1
Pygments==2.15.1
PyMCubes==0.1.4
pyparsing==3.0.9
pypng==0.0.20
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyquaternion==0.9.9
pyrsistent==0.19.3
pytest==7.3.1
python-dateutil==2.8.2
pytorch-lightning==1.3.5
pytorch-msssim==1.0.0
pytz==2023.3
PyWavelets==1.3.0
PyYAML==5.4.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
regex==2024.4.16
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.4.1
rsa==4.9
safetensors==0.3.1
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.2
Shapely==1.8.5
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.0
soupsieve==2.4.1
spatial-correlation-sampler==0.5.0
spconv-cu113==2.3.6
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
terminado==0.17.1
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.9.2
tinycss2==1.2.1
tomli==2.0.1
torch==1.10.1
torch-scatter==2.1.1
torchaudio==0.10.1
torchmetrics==0.5.1
torchvision==0.11.2
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
traits==6.4.1
traitsui==8.0.0
trimesh==2.35.39
typed-ast==1.5.4
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1685704949284/work
urllib3==1.26.16
vtk==9.1.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.2
Werkzeug==2.2.3
widgetsnbextension==4.0.7
wslink==1.11.1
yacs==0.1.8
yapf==0.33.0
yarl==1.9.2
zipp==3.15.0
